Factor out rpath splitting
With repo remapping in play, since Bazel 6 and bzlmod, we need to
consider the repository and path components of an rpath separately. This
is particularly relevant for the manifest based runfiles strategy where
we need to look up runfiles paths in a key-value mapping. If we were to
use simple strings as keys in this mapping then a runfiles lookup that
involves remapping would always require an extra allocation to construct
the new runfiles path with the remapped repository. This is unnecessary,
if we simply consider the repository and path component separately, then
we can avoid this extra allocation.