Store structured module arguments
Zig's module system is changing drastically between releases. However,
the direction seems to be a system where each module is represented on
the command-line by a either multiple arguments in a specific order
(0.12.0) or a single flag with a complex structure (0.11.0). Neither of
those lends itself to tracking parts of the command-line separately in
`depset` values to later on pass to `args.add_all`.

Instead, we track modules in a `depset` of `struct` that carries all
relevant information for each module, and later apply `args.add_all`
using the `map_each` parameter to render the flags for each module.