Rewrite tests for zig_module_info
While using depsets and postorder ordering, I realized that _mock_args
implementation was different than ctx.actions.args especially in the
case where depsets are using postorder.

In this commit, I suggest removing the need for the mock and instead
test zig_module_specifications with the real implementation of
ctx.actions.args using custom rules and a diff_test.

This guarantees that we are testing the exact same order of args
as they will appear in the command line when ctx.actions.run is
performed.